round of fixes:
- Crash on empty frames — driver returns zero-length buffers at stream start → cannot reshape array of size 0: get_frame/stream skip short frames until a full one
- UI sliders never synced — API keyed controls by display name ("Gain"), JS looked up gain → all missing; sliders stuck on wrong hardcoded ranges
- Lossy "lossless" capture — capture-while-streaming re-encoded the q=85 JPEG into PNG/TIFF/FITS
- SER WB flicker — per-frame gray-world gains varied frame-to-frame, ruining stacking
- SER 16-bit endianness: Explicit little-endian (<u2) writes; corrected header comment
- UI record bugs — read nonexistent data.filename, duplicate element id, no UI reset on auto-stop

+++ camera/export.py
@@ -176,7 +176,10 @@ class SERWriter:
         if self._bits == 8:
             data = frame.astype(np.uint8)
         else:
-            data = frame.astype(np.uint16)
+            # Write little-endian explicitly so output is correct regardless of
+            # host byte order; matches the LittleEndian header flag (see
+            # _make_header).
+            data = frame.astype("<u2")
 
         self._file.write(data.tobytes())
         self._frame_count += 1
@@ -195,7 +198,10 @@ class SERWriter:
             b"LUCAM-RECORDER"              # file ID (14 bytes)
             + struct.pack("<i", 0)         # LuID
             + struct.pack("<i", color_id)  # ColorID
-            + struct.pack("<i", 0)         # LittleEndian (0 = big-endian pixel order per spec)
+            # LittleEndian flag for 16-bit pixel data. The SER ecosystem
+            # (SER Player, PIPP) de-facto reads 0 as little-endian, and we
+            # write little-endian data (see write_frame), so 0 is correct.
+            + struct.pack("<i", 0)
             + struct.pack("<i", self._width)
             + struct.pack("<i", self._height)
             + struct.pack("<i", self._bits)
